+++ core/src/main/java/com/graphhopper/routing/ev/IntEncodedValueImpl.java
@@ -189,6 +189,9 @@ public class IntEncodedValueImpl implements IntEncodedValue {
 
     @Override
     public final int getInt(boolean reverse, int edgeId, EdgeIntAccess edgeIntAccess) {
+        assert fwdShift >= 0 : "incorrect shift " + fwdShift + " for " + getName();
+        assert bits > 0 : "incorrect bits " + bits + " for " + getName();
+
         int flags;
         // if we do not store both directions ignore reverse == true for convenient reading
         if (storeTwoDirections && reverse) {