+++ core/src/main/java/com/graphhopper/storage/BaseGraphNodesAndEdges.java
@@ -18,6 +18,7 @@
 
 package com.graphhopper.storage;
 
+import com.graphhopper.routing.ev.EdgeIntAccess;
 import com.graphhopper.util.*;
 import com.graphhopper.util.shapes.BBox;
 
@@ -30,7 +31,7 @@ import static com.graphhopper.util.Helper.nf;
  * Underlying storage for nodes and edges of {@link BaseGraph}. Nodes and edges are stored using two {@link DataAccess}
  * instances. Nodes and edges are simply stored sequentially, see the memory layout in the constructor.
  */
-class BaseGraphNodesAndEdges {
+class BaseGraphNodesAndEdges implements EdgeIntAccess {
     // Currently distances are stored as 4 byte integers. using a conversion factor of 1000 the minimum distance
     // that is not considered zero is 0.0005m (=0.5mm) and the maximum distance per edge is about 2.147.483m=2147km.
     // See OSMReader.addEdge and #1871.
@@ -45,8 +46,8 @@ class BaseGraphNodesAndEdges {
 
     // edges
     private final DataAccess edges;
-    private final int E_NODEA, E_NODEB, E_LINKA, E_LINKB, E_FLAGS, E_DIST, E_GEO_1, E_GEO_2, E_KV;
-    private final int intsForFlags;
+    private final int E_NODEA, E_NODEB, E_LINKA, E_LINKB, E_DIST, E_KV, E_FLAGS, E_GEO;
+    private final int bytesForFlags;
     private int edgeEntryBytes;
     private int edgeCount;
 
@@ -59,10 +60,10 @@ class BaseGraphNodesAndEdges {
     public final BBox bounds;
     private boolean frozen;
 
-    public BaseGraphNodesAndEdges(Directory dir, int intsForFlags, boolean withElevation, boolean withTurnCosts, int segmentSize) {
+    public BaseGraphNodesAndEdges(Directory dir, boolean withElevation, boolean withTurnCosts, int segmentSize, int bytesForFlags) {
         nodes = dir.create("nodes", dir.getDefaultType("nodes", true), segmentSize);
-        edges = dir.create("edges", dir.getDefaultType("edges", true), segmentSize);
-        this.intsForFlags = intsForFlags;
+        edges = dir.create("edges", dir.getDefaultType("edges", false), segmentSize);
+        this.bytesForFlags = bytesForFlags;
         this.withTurnCosts = withTurnCosts;
         this.withElevation = withElevation;
         bounds = BBox.createInverse(withElevation);
@@ -80,12 +81,11 @@ class BaseGraphNodesAndEdges {
         E_NODEB = 4;
         E_LINKA = 8;
         E_LINKB = 12;
-        E_FLAGS = 16;
-        E_DIST = E_FLAGS + intsForFlags * 4;
-        E_GEO_1 = E_DIST + 4;
-        E_GEO_2 = E_GEO_1 + 4;
-        E_KV = E_GEO_2 + 4;
-        edgeEntryBytes = E_KV + 4;
+        E_DIST = 16;
+        E_KV = 20;
+        E_FLAGS = 24;
+        E_GEO = E_FLAGS + bytesForFlags + 4;
+        edgeEntryBytes = E_GEO + 5;
     }
 
     public void create(long initSize) {
@@ -160,8 +160,12 @@ class BaseGraphNodesAndEdges {
         return edgeCount;
     }
 
-    public int getIntsForFlags() {
-        return intsForFlags;
+    IntsRef createEdgeFlags() {
+        return new IntsRef((int) Math.ceil((double) getBytesForFlags() / 4));
+    }
+
+    public int getBytesForFlags() {
+        return bytesForFlags;
     }
 
     public boolean withElevation() {
@@ -243,21 +247,59 @@ class BaseGraphNodesAndEdges {
     public void readFlags(long edgePointer, IntsRef edgeFlags) {
         int size = edgeFlags.ints.length;
         for (int i = 0; i < size; ++i)
-            edgeFlags.ints[i] = getFlagInt(edgePointer, i);
+            edgeFlags.ints[i] = getFlagInt(edgePointer, i * 4);
     }
 
     public void writeFlags(long edgePointer, IntsRef edgeFlags) {
         int size = edgeFlags.ints.length;
         for (int i = 0; i < size; ++i)
-            setFlagInt(edgePointer, i, edgeFlags.ints[i]);
+            setFlagInt(edgePointer, i * 4, edgeFlags.ints[i]);
     }
 
-    public int getFlagInt(long edgePointer, int index) {
-        return edges.getInt(edgePointer + E_FLAGS + index * 4);
+    private int getFlagInt(long edgePointer, int byteOffset) {
+        if (byteOffset >= bytesForFlags)
+            throw new IllegalArgumentException("too large byteOffset " + byteOffset + " vs " + bytesForFlags);
+        edgePointer += byteOffset;
+        if (byteOffset + 3 == bytesForFlags) {
+            return (edges.getShort(edgePointer + E_FLAGS) << 8) & 0x00FF_FFFF | edges.getByte(edgePointer + E_FLAGS + 2) & 0xFF;
+        } else if (byteOffset + 2 == bytesForFlags) {
+            return edges.getShort(edgePointer + E_FLAGS) & 0xFFFF;
+        } else if (byteOffset + 1 == bytesForFlags) {
+            return edges.getByte(edgePointer + E_FLAGS) & 0xFF;
+        }
+        return edges.getInt(edgePointer + E_FLAGS);
     }
 
-    public void setFlagInt(long edgePointer, int index, int value) {
-        edges.setInt(edgePointer + E_FLAGS + index * 4, value);
+    private void setFlagInt(long edgePointer, int byteOffset, int value) {
+        if (byteOffset >= bytesForFlags)
+            throw new IllegalArgumentException("too large byteOffset " + byteOffset + " vs " + bytesForFlags);
+        edgePointer += byteOffset;
+        if (byteOffset + 3 == bytesForFlags) {
+            if ((value & 0xFF00_0000) != 0)
+                throw new IllegalArgumentException("value at byteOffset " + byteOffset + " must not have the highest byte set but was " + value);
+            edges.setShort(edgePointer + E_FLAGS, (short) (value >> 8));
+            edges.setByte(edgePointer + E_FLAGS + 2, (byte) value);
+        } else if (byteOffset + 2 == bytesForFlags) {
+            if ((value & 0xFFFF_0000) != 0)
+                throw new IllegalArgumentException("value at byteOffset " + byteOffset + " must not have the 2 highest bytes set but was " + value);
+            edges.setShort(edgePointer + E_FLAGS, (short) value);
+        } else if (byteOffset + 1 == bytesForFlags) {
+            if ((value & 0xFFFF_FF00) != 0)
+                throw new IllegalArgumentException("value at byteOffset " + byteOffset + " must not have the 3 highest bytes set but was " + value);
+            edges.setByte(edgePointer + E_FLAGS, (byte) value);
+        } else {
+            edges.setInt(edgePointer + E_FLAGS, value);
+        }
+    }
+
+    @Override
+    public int getInt(int edgeId, int index) {
+        return getFlagInt(toEdgePointer(edgeId), index * 4);
+    }
+
+    @Override
+    public void setInt(int edgeId, int index, int value) {
+        setFlagInt(toEdgePointer(edgeId), index * 4, value);
     }
 
     public void setNodeA(long edgePointer, int nodeA) {
@@ -281,10 +323,12 @@ class BaseGraphNodesAndEdges {
     }
 
     public void setGeoRef(long edgePointer, long geoRef) {
-        int geo1 = BitUtil.LITTLE.getIntLow(geoRef);
-        int geo2 = BitUtil.LITTLE.getIntHigh(geoRef);
-        edges.setInt(edgePointer + E_GEO_1, geo1);
-        edges.setInt(edgePointer + E_GEO_2, geo2);
+        if ((geoRef & 0xFFFF_FF00_0000_0000L) != 0)
+            throw new IllegalArgumentException("geoRef is too large " + geoRef);
+
+        byte[] bytes = new byte[5];
+        BitUtil.LITTLE.fromULong5(bytes, geoRef, 0);
+        edges.setBytes(edgePointer + E_GEO, bytes, bytes.length);
     }
 
     public void setKeyValuesRef(long edgePointer, int nameRef) {
@@ -314,10 +358,9 @@ class BaseGraphNodesAndEdges {
     }
 
     public long getGeoRef(long edgePointer) {
-        return BitUtil.LITTLE.toLong(
-                edges.getInt(edgePointer + E_GEO_1),
-                edges.getInt(edgePointer + E_GEO_2)
-        );
+        byte[] bytes = new byte[5];
+        edges.getBytes(edgePointer + E_GEO, bytes, bytes.length);
+        return BitUtil.LITTLE.toULong5(bytes, 0);
     }
 
     public int getKeyValuesRef(long edgePointer) {
@@ -387,16 +430,16 @@ class BaseGraphNodesAndEdges {
         System.out.println("edges:");
         String formatEdges = "%12s | %12s | %12s | %12s | %12s | %12s | %12s \n";
         System.out.format(Locale.ROOT, formatEdges, "#", "E_NODEA", "E_NODEB", "E_LINKA", "E_LINKB", "E_FLAGS", "E_DIST");
-        IntsRef intsRef = new IntsRef(intsForFlags);
+        IntsRef edgeFlags = createEdgeFlags();
         for (int i = 0; i < Math.min(edgeCount, printMax); ++i) {
             long edgePointer = toEdgePointer(i);
-            readFlags(edgePointer, intsRef);
+            readFlags(edgePointer, edgeFlags);
             System.out.format(Locale.ROOT, formatEdges, i,
                     getNodeA(edgePointer),
                     getNodeB(edgePointer),
                     getLinkA(edgePointer),
                     getLinkB(edgePointer),
-                    intsRef,
+                    edgeFlags,
                     getDist(edgePointer));
         }
         if (edgeCount > printMax) {