+++ core/src/test/java/com/graphhopper/util/InstructionListTest.java
@@ -35,13 +35,10 @@ import com.graphhopper.storage.NodeAccess;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Locale;
+import java.util.*;
 
-import static com.graphhopper.search.KVStorage.KeyValue.STREET_NAME;
-import static com.graphhopper.search.KVStorage.KeyValue.createKV;
+import static com.graphhopper.search.KVStorage.KValue;
+import static com.graphhopper.util.Parameters.Details.STREET_NAME;
 import static org.junit.jupiter.api.Assertions.*;
 
 /**
@@ -93,30 +90,30 @@ public class InstructionListTest {
         na.setNode(6, 1.0, 1.0);
         na.setNode(7, 1.0, 1.1);
         na.setNode(8, 1.0, 1.2);
-        g.edge(0, 1).setDistance(10000).set(speedEnc, 60, 60).setKeyValues(createKV(STREET_NAME, "0-1"));
-        g.edge(1, 2).setDistance(11000).set(speedEnc, 60, 60).setKeyValues(createKV(STREET_NAME, "1-2"));
+        g.edge(0, 1).setDistance(10000).set(speedEnc, 60, 60).setKeyValues(Map.of(STREET_NAME, new KValue("0-1")));
+        g.edge(1, 2).setDistance(11000).set(speedEnc, 60, 60).setKeyValues(Map.of(STREET_NAME, new KValue("1-2")));
 
         g.edge(0, 3).setDistance(11000).set(speedEnc, 60, 60);
-        g.edge(1, 4).setDistance(10000).set(speedEnc, 60, 60).setKeyValues(createKV(STREET_NAME, "1-4"));
-        g.edge(2, 5).setDistance(11000).set(speedEnc, 60, 60).setKeyValues(createKV(STREET_NAME, "5-2"));
+        g.edge(1, 4).setDistance(10000).set(speedEnc, 60, 60).setKeyValues(Map.of(STREET_NAME, new KValue("1-4")));
+        g.edge(2, 5).setDistance(11000).set(speedEnc, 60, 60).setKeyValues(Map.of(STREET_NAME, new KValue("5-2")));
 
-        g.edge(3, 6).setDistance(11000).set(speedEnc, 60, 60).setKeyValues(createKV(STREET_NAME, "3-6"));
-        g.edge(4, 7).setDistance(10000).set(speedEnc, 60, 60).setKeyValues(createKV(STREET_NAME, "4-7"));
-        g.edge(5, 8).setDistance(10000).set(speedEnc, 60, 60).setKeyValues(createKV(STREET_NAME, "5-8"));
+        g.edge(3, 6).setDistance(11000).set(speedEnc, 60, 60).setKeyValues(Map.of(STREET_NAME, new KValue("3-6")));
+        g.edge(4, 7).setDistance(10000).set(speedEnc, 60, 60).setKeyValues(Map.of(STREET_NAME, new KValue("4-7")));
+        g.edge(5, 8).setDistance(10000).set(speedEnc, 60, 60).setKeyValues(Map.of(STREET_NAME, new KValue("5-8")));
 
-        g.edge(6, 7).setDistance(11000).set(speedEnc, 60, 60).setKeyValues(createKV(STREET_NAME, "6-7"));
+        g.edge(6, 7).setDistance(11000).set(speedEnc, 60, 60).setKeyValues(Map.of(STREET_NAME, new KValue("6-7")));
         EdgeIteratorState iter = g.edge(7, 8).setDistance(10000).set(speedEnc, 60, 60);
         PointList list = new PointList();
         list.add(1.0, 1.15);
         list.add(1.0, 1.16);
         iter.setWayGeometry(list);
-        iter.setKeyValues(createKV(STREET_NAME, "7-8"));
+        iter.setKeyValues(Map.of(STREET_NAME, new KValue("7-8")));
         // missing edge name
         g.edge(9, 10).setDistance(10000).set(speedEnc, 60, 60);
         EdgeIteratorState iter2 = g.edge(8, 9).setDistance(20000).set(speedEnc, 60, 60);
         list.clear();
         list.add(1.0, 1.3);
-        iter2.setKeyValues(createKV(STREET_NAME, "8-9"));
+        iter2.setKeyValues(Map.of(STREET_NAME, new KValue("8-9")));
         iter2.setWayGeometry(list);
         return g;
     }
@@ -185,11 +182,11 @@ public class InstructionListTest {
         na.setNode(3, 10.0, 10.08);
         na.setNode(4, 10.1, 10.10);
         na.setNode(5, 10.2, 10.13);
-        g.edge(3, 4).setDistance(100).set(speedEnc, 60, 60).setKeyValues(createKV(STREET_NAME, "3-4"));
-        g.edge(4, 5).setDistance(100).set(speedEnc, 60, 60).setKeyValues(createKV(STREET_NAME, "4-5"));
+        g.edge(3, 4).setDistance(100).set(speedEnc, 60, 60).setKeyValues(Map.of(STREET_NAME, new KValue("3-4")));
+        g.edge(4, 5).setDistance(100).set(speedEnc, 60, 60).setKeyValues(Map.of(STREET_NAME, new KValue("4-5")));
 
         EdgeIteratorState iter = g.edge(2, 4).setDistance(100).set(speedEnc, 60, 60);
-        iter.setKeyValues(createKV(STREET_NAME, "2-4"));
+        iter.setKeyValues(Map.of(STREET_NAME, new KValue("2-4")));
         PointList list = new PointList();
         list.add(10.20, 10.05);
         iter.setWayGeometry(list);
@@ -224,11 +221,11 @@ public class InstructionListTest {
         na.setNode(3, 10.0, 10.05);
         na.setNode(4, 10.1, 10.10);
         na.setNode(5, 10.2, 10.15);
-        g.edge(3, 4).setDistance(100).set(speedEnc, 60, 60).setKeyValues(createKV(STREET_NAME, "street"));
-        g.edge(4, 5).setDistance(100).set(speedEnc, 60, 60).setKeyValues(createKV(STREET_NAME, "4-5"));
+        g.edge(3, 4).setDistance(100).set(speedEnc, 60, 60).setKeyValues(Map.of(STREET_NAME, new KValue("street")));
+        g.edge(4, 5).setDistance(100).set(speedEnc, 60, 60).setKeyValues(Map.of(STREET_NAME, new KValue("4-5")));
 
         EdgeIteratorState iter = g.edge(2, 4).setDistance(100).set(speedEnc, 60, 60);
-        iter.setKeyValues(createKV(STREET_NAME, "street"));
+        iter.setKeyValues(Map.of(STREET_NAME, new KValue("street")));
         PointList list = new PointList();
         list.add(10.20, 10.05);
         iter.setWayGeometry(list);
@@ -324,9 +321,9 @@ public class InstructionListTest {
         g.edge(2, 3).setDistance(20).set(speedEnc, 18, 18);
         g.edge(2, 4).setDistance(20).set(speedEnc, 4, 4);
 
-        g.edge(1, 2).set(rcEV, RoadClass.RESIDENTIAL).setKeyValues(createKV(STREET_NAME, "pfarr"));
-        g.edge(2, 3).set(rcEV, RoadClass.RESIDENTIAL).setKeyValues(createKV(STREET_NAME, "pfarr"));
-        g.edge(2, 4).set(rcEV, RoadClass.PEDESTRIAN).setKeyValues(createKV(STREET_NAME, "markt"));
+        g.edge(1, 2).set(rcEV, RoadClass.RESIDENTIAL).setKeyValues(Map.of(STREET_NAME, new KValue("pfarr")));
+        g.edge(2, 3).set(rcEV, RoadClass.RESIDENTIAL).setKeyValues(Map.of(STREET_NAME, new KValue("pfarr")));
+        g.edge(2, 4).set(rcEV, RoadClass.PEDESTRIAN).setKeyValues(Map.of(STREET_NAME, new KValue("markt")));
 
         Weighting weighting = new SpeedWeighting(speedEnc);
         Path p = new Dijkstra(g, weighting, tMode).calcPath(1, 3);
@@ -399,13 +396,13 @@ public class InstructionListTest {
 
         // default is priority=0 so set it to 1
         g.edge(1, 2).setDistance(20).set(speedEnc, 5).
-                setKeyValues(createKV(STREET_NAME, "myroad"));
+                setKeyValues(Map.of(STREET_NAME, new KValue("myroad")));
         g.edge(2, 3).setDistance(20).set(speedEnc, 5).
-                setKeyValues(createKV(STREET_NAME, "myroad"));
+                setKeyValues(Map.of(STREET_NAME, new KValue("myroad")));
         PointList pointList = new PointList();
         pointList.add(43.729627, 7.41749);
         g.edge(2, 4).setDistance(20).set(speedEnc, 5).
-                setKeyValues(createKV(STREET_NAME, "myroad")).setWayGeometry(pointList);
+                setKeyValues(Map.of(STREET_NAME, new KValue("myroad"))).setWayGeometry(pointList);
 
         Weighting weighting = new SpeedWeighting(speedEnc);
         Path p = new Dijkstra(g, weighting, tMode).calcPath(4, 3);
@@ -495,12 +492,12 @@ public class InstructionListTest {
         na.setNode(6, 15.1, 10.1);
         na.setNode(7, 15.1, 9.8);
 
-        g.edge(1, 2).setDistance(10000).set(speedEnc, 60, 60).setKeyValues(createKV(STREET_NAME, "1-2"));
-        g.edge(2, 3).setDistance(10000).set(speedEnc, 60, 60).setKeyValues(createKV(STREET_NAME, "2-3"));
-        g.edge(2, 6).setDistance(10000).set(speedEnc, 60, 60).setKeyValues(createKV(STREET_NAME, "2-6"));
-        g.edge(3, 4).setDistance(10000).set(speedEnc, 60, 60).setKeyValues(createKV(STREET_NAME, "3-4")).setWayGeometry(waypoint);
-        g.edge(3, 7).setDistance(10000).set(speedEnc, 60, 60).setKeyValues(createKV(STREET_NAME, "3-7"));
-        g.edge(4, 5).setDistance(10000).set(speedEnc, 60, 60).setKeyValues(createKV(STREET_NAME, "4-5"));
+        g.edge(1, 2).setDistance(10000).set(speedEnc, 60, 60).setKeyValues(Map.of(STREET_NAME, new KValue("1-2")));
+        g.edge(2, 3).setDistance(10000).set(speedEnc, 60, 60).setKeyValues(Map.of(STREET_NAME, new KValue("2-3")));
+        g.edge(2, 6).setDistance(10000).set(speedEnc, 60, 60).setKeyValues(Map.of(STREET_NAME, new KValue("2-6")));
+        g.edge(3, 4).setDistance(10000).set(speedEnc, 60, 60).setKeyValues(Map.of(STREET_NAME, new KValue("3-4"))).setWayGeometry(waypoint);
+        g.edge(3, 7).setDistance(10000).set(speedEnc, 60, 60).setKeyValues(Map.of(STREET_NAME, new KValue("3-7")));
+        g.edge(4, 5).setDistance(10000).set(speedEnc, 60, 60).setKeyValues(Map.of(STREET_NAME, new KValue("4-5")));
 
         Weighting weighting = new SpeedWeighting(speedEnc);
         Path p = new Dijkstra(g, weighting, tMode).calcPath(1, 5);
@@ -544,11 +541,11 @@ public class InstructionListTest {
 
         PointList list = new PointList();
         list.add(43.62549, -79.714292);
-        g.edge(1, 2).setKeyValues(createKV(STREET_NAME, "main")).setWayGeometry(list).
+        g.edge(1, 2).setKeyValues(Map.of(STREET_NAME, new KValue("main"))).setWayGeometry(list).
                 setDistance(110).set(roadsSpeedEnc, 50, 50).set(lanesEnc, 2);
-        g.edge(2, 3).setKeyValues(createKV(STREET_NAME, "main")).
+        g.edge(2, 3).setKeyValues(Map.of(STREET_NAME, new KValue("main"))).
                 setDistance(110).set(roadsSpeedEnc, 50, 50).set(lanesEnc, 3);
-        g.edge(2, 4).setKeyValues(createKV(STREET_NAME, "main")).
+        g.edge(2, 4).setKeyValues(Map.of(STREET_NAME, new KValue("main"))).
                 setDistance(80).set(roadsSpeedEnc, 50, 50).set(lanesEnc, 5);
 
         Weighting weighting = new SpeedWeighting(roadsSpeedEnc);